refactor: explain lifi integration
