+++ src/app/[locale]/admin/users/page.tsx
@@ -1,5 +1,5 @@
 import { setRequestLocale } from 'next-intl/server'
-import AdminUsersTable from '@/app/[locale]/admin/(users)/_components/AdminUsersTable'
+import AdminUsersTable from '@/app/[locale]/admin/users/_components/AdminUsersTable'
 
 export default async function AdminUsersPage({ params }: PageProps<'/[locale]/admin/users'>) {
   const { locale } = await params
